+++ src/encode.hpp
@@ -42,8 +42,8 @@ THE SOFTWARE.
 using namespace std;
 
 constexpr uint8_t CODE_VALUE_BITS = 32;
-constexpr uint64_t MAX_CODE = (1ULL<<CODE_VALUE_BITS)-1;
-constexpr uint64_t ONE_FOURTH = (MAX_CODE+1ULL)/4;
+constexpr uint64_t MAX_CODE = (((uint64_t)1) << CODE_VALUE_BITS)-1;
+constexpr uint64_t ONE_FOURTH = (MAX_CODE + ((uint64_t)1))/4;
 constexpr uint64_t ONE_HALF = ONE_FOURTH*2;
 constexpr uint64_t THREE_FOURTHS = ONE_FOURTH*3;
